Unfork startup files
to reduce conflicts in the ginkgo merge.
