+++ cms/startup.py
@@ -8,7 +8,6 @@ from django.conf import settings
 
 settings.INSTALLED_APPS  # pylint: disable=pointless-statement
 
-from openedx.core.lib.django_startup import add_mimetypes
 from openedx.core.lib.django_startup import autostartup
 import django
 from openedx.core.djangoapps.monkey_patch import django_db_models_options
@@ -48,3 +47,19 @@ def run():
 
     # validate configurations on startup
     validate_cms_config(settings)
+
+
+def add_mimetypes():
+    """
+    Add extra mimetypes. Used in xblock_resource.
+
+    If you add a mimetype here, be sure to also add it in lms/startup.py.
+    """
+    import mimetypes
+
+    mimetypes.add_type('application/vnd.ms-fontobject', '.eot')
+    mimetypes.add_type('application/x-font-opentype', '.otf')
+    mimetypes.add_type('application/x-font-ttf', '.ttf')
+    mimetypes.add_type('application/font-woff', '.woff')
+    for extension, mimetype in settings.EXTRA_MIMETYPES.iteritems():
+        mimetypes.add_type(mimetype, extension)
