+++ openedx/core/lib/django_startup.py
@@ -1,23 +1,9 @@
 """
-Extend startup behavior shared between LMS and CMS
+Automatic execution of startup modules in Django apps.
 """
 
 from importlib import import_module
 from django.conf import settings
-
-
-def add_mimetypes():
-    """
-    Add extra mimetype mappings
-    """
-    import mimetypes
-
-    mimetypes.add_type('application/vnd.ms-fontobject', '.eot')
-    mimetypes.add_type('application/x-font-opentype', '.otf')
-    mimetypes.add_type('application/x-font-ttf', '.ttf')
-    mimetypes.add_type('application/font-woff', '.woff')
-    for extension, mimetype in settings.EXTRA_MIMETYPES.iteritems():
-        mimetypes.add_type(mimetype, extension)
 
 
 def autostartup():
